fix: headlines should be the same on all native platforms
