+++ packages/article-magazine-standard/src/styles/index.js
@@ -1,8 +1,21 @@
 import { StyleSheet } from "react-native";
+import { spacing } from "@times-components/styleguide";
 import nativeStyles from "./native";
 
+const androidStyles = {
+  ...nativeStyles,
+  flags: {
+    ...nativeStyles.flags,
+    marginTop: spacing(2)
+  },
+  metaTablet: {
+    ...nativeStyles.metaTablet,
+    paddingTop: "auto"
+  }
+};
+
 const styles = StyleSheet.create({
-  ...nativeStyles
+  ...androidStyles
 });
 
 export default styles;
